Add reference to SQLite URI Filenames in Connect Options
In many ORMs and connectors the SQLite database is configured through functions such as `.create_if_missing(true)`. SeaORM does not provide this to prevent leaking abstractions that make no sense for Postgres or MySQL. This requires the user to know about 

All of the existing examples in SeaORM use `sqlite::memory:` and a user making an intuitive guess of `sqlite:data.db` or `sqlite://data.db` whilst trying for a persistent SQLite database will be met with errors regarding not finding / opening the given file.

I propose adding a brief mention to the SeaORM documentation directing the users to SQLite's URI filenames and providing the specific example of `sqlite://data.db?mode=rwc` which would likely cover the majority of use cases. Guidance on achieving clearer writing or finding a better home for this content in the documentation is welcomed.